requirements.txt
numpy
opencv-python
matplotlib
tensorflow
keras
mtcnn
pillow